+++ resources/js/hiroapp/components/Footer.jsx
